// web/src/api/generated/models/salesSetDraftItemQuantityCommand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */

export interface SalesSetDraftItemQuantityCommand {
  quantity?: number;
  request_id?: string;
}