// web/src/api/generated/models/salesSetDraftItemModifiersCommand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */

export interface SalesSetDraftItemModifiersCommand {
  modifier_option_ids?: string[];
  request_id?: string;
}