// web/src/api/generated/models/salesSetDraftItemNoteCommand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */

export interface SalesSetDraftItemNoteCommand {
  preparation_note?: string;
  request_id?: string;
}